Fixes figure resize behavior in HG2
The way the program was designed, there are many components that listen to
changes to the figure position and react accordingly. There are several
points at which this behavior becomes circular (in HG2) and also explicit
drawnows had to be added in order for the Position property of the figure
to actually be updated and then used to calculate new control positions. To
simplify things, all listeners are turned off by DENSEanalysis.m and then
a single ResizeFcn is assigned to the main figure which then performs all
of the redraw() commands in the appropriate order. This explicit redrawing
forces all components to be updated in the proper order and ensures that the
display is appropriate.