Adds ability to export transmural strains via MAT file (Closes #4)
DENSEanalysis was only able to compute transmural strain values when the
user exported the results as Excel files. This limits the functionality
because they must then be re-converted to MAT files to be processed.

All transmural strain data is now written to the 'TransmuralStrainInfo'
variable in the returned MAT file.